Created the women's perfumes component and page.
Created the women's perfumes component and page. Added the path to the page through routes and path which I could then add the path to the page in the navbar so you can navigate to it. Also added the router outlet to allow the content of the page to be displayed