Created so that the perfumes for women are printed with data and styling
Created an interface for my array to describe the data structure of the project. Also deleted the file I used to display the products in before and used a service instead of other components accessing the data (used the same functions and data). Created a component that loops through all the products and creates a div for each to position them correctly on the page. In the loop, the products are sent to another component, which in turn prints out all the data for the products. Also added styling for them so that the products look exactly the same as for the reactjs application. Also created a folder in assets with images and added the path to the function that loops images