Has creates the functionality for the product to end up in the shopping cart.
Has creates the functionality for the product to end up in the shopping cart. Started by creating components for the shopping cart. In part, a service was created where addProductToCart was created and pushes the product that is chosen to be added to the shopping cart to CartItem and a getproductCart that returns the cart. A component was also created that used setProductCart from cartService to getProductCart. As well as a component called cartitem and one called cartproduct which puts the product from cartItem into an array which is then used in the component cart which prints the product chosen to be added to the shopping cart. Also added styling for the products in the cart.