Created a button on the products in the cart that deletes a product in the cart.
Created a button on the products in the cart that deletes a product in the cart. Created a function (removeProductFromCart) that was added to the cartService component that adds id to the product when the user clicks the button that deletes the product and in the cart.component file the function runs and updates the products in the shopping cart which makes the product gone. Also added styling for the button so it looks the same as the reactjs project