Fixed so that the products that are created are always the same through localStorage
Fixed so that the products that are created are always the same through localStorage so that it does not affect the measurements. Created a function in product.Service retrieves the products from the generatePerfume() function that adds the data from the different functions that create the data to localStorage through JSON string. Then a function that retrieves data in localStorage and returns a JSON string that is then used in the component that prints the products. In the constructor of that component, a function was also added that checks if localStorage has not retrieved the data, if it has not, the function that saves localStorage in the product.service component is called