Created total price in cart
Created functionality to add up the products in the shopping cart costs to see the total price. Added a function in cartProduct that loops through all products that are in cartItem and adds their prices to variable cartTotalPrice which is then returned. Then prints the total price on the shopping cart page. Also added styling.