Created the same functionality as for the women's perfumes but for the men's perfumes
Created the same functionality as for the women's perfumes but for the men's perfumes. Created a new service where the data for the products is created fixed with men information. Also created a new model for men perfume. Created a component that loops through all the products and sends that data to app-menPerfume. App-menPerfume then prints the data correctly. Uses the same functions as addProductToCart in the cart components. Finally I added all the pictures for but the perfumes