This was done to create two-way data binding for the products to end up in the shopping cart.
This was done to create two-way data binding for the products to end up in the shopping cart. The first thing that was done was to implement the FormsModule in app.module. Next, NgModel was added to both the components that print the products for women's and men's perfumes. The quantity that the user chooses to add is sent with the addToCartProduct function. In addToCartProduct, the quantity of products is then added to be then added to cartItem and then printed on the shopping cart page