Added so that the product is also added to the cart through two way data binding
Added so that the product is also added to the cart through two way data binding by using ngModelChange in the input when the user selects the number of products. Did this for both the women's and men's perfumes. Due to the fact that the data is updated when the user selects the number, the alert came up when I was about to fill in the number, so I chose to remove the alert, but the if clause is still there, so you can still only add a product if the user has filled in more than zero products . Also needed to increase the height of the product dive so that the input would be placed well