Added functionality to check if the product is already in the shopping cart
Has added functionality to check if the product is already in the shopping cart and if it is it just updates quantity instead of adding the product again. Through Find it is checked if ID is already in Cart Item, which is an array containing all products already in the shopping cart. Everything is checked through an IF SATS and if the product is already in the shopping cart, Quantity is on with an otherwise pusha product and quantity to Cartitem and printed